Move classes to internal package and update manifest
