+++ addons/binding/org.openhab.binding.neeo/src/main/java/org/openhab/binding/neeo/internal/NeeoBrainConfig.java
@@ -10,7 +10,7 @@ package org.openhab.binding.neeo.internal;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.neeo.handler.NeeoBrainHandler;
+import org.openhab.binding.neeo.internal.handler.NeeoBrainHandler;
 
 /**
  * The configuration class a neeo brain and used by {@link NeeoBrainHandler}
