+++ addons/binding/org.openhab.binding.neeo/src/main/java/org/openhab/binding/neeo/internal/NeeoDeviceConfig.java
@@ -10,7 +10,7 @@ package org.openhab.binding.neeo.internal;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.neeo.handler.NeeoDeviceHandler;
+import org.openhab.binding.neeo.internal.handler.NeeoDeviceHandler;
 
 /**
  * THe configuration class for the device used by {@link NeeoDeviceHandler}
