+++ addons/binding/org.openhab.binding.neeo/src/main/java/org/openhab/binding/neeo/internal/NeeoRoomConfig.java
@@ -10,7 +10,7 @@ package org.openhab.binding.neeo.internal;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.neeo.handler.NeeoRoomHandler;
+import org.openhab.binding.neeo.internal.handler.NeeoRoomHandler;
 
 /**
  * THe configuration class for the room used by {@link NeeoRoomHandler}
