+++ addons/binding/org.openhab.binding.neeo/src/main/java/org/openhab/binding/neeo/internal/discovery/NeeoRoomDiscoveryService.java
@@ -21,10 +21,10 @@ import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
 import org.eclipse.smarthome.core.thing.Bridge;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
-import org.openhab.binding.neeo.handler.NeeoBrainHandler;
 import org.openhab.binding.neeo.internal.NeeoBrainApi;
 import org.openhab.binding.neeo.internal.NeeoBrainConfig;
 import org.openhab.binding.neeo.internal.NeeoConstants;
+import org.openhab.binding.neeo.internal.handler.NeeoBrainHandler;
 import org.openhab.binding.neeo.internal.models.NeeoBrain;
 import org.openhab.binding.neeo.internal.models.NeeoRoom;
 import org.slf4j.Logger;
